Adding package scripts
Script to create a shell bundle file that handles the whole installation
on a fresh Ubuntu 15 machine.

You have to run the "create-bundle.sh" somewhere to generate the
"ignite-bundle.sh" file, which is the installer script inclusive of the
code part of this app.

Once you have the "ignite-bundle.sh", you can distribute that, and it
has been tested to work on Ubuntu 15 server.